+++ ErsatzTV.Application/Health/Queries/GetZeroDurationFiles.cs
@@ -1,0 +1,3 @@
+namespace ErsatzTV.Application.Health;
+
+public record GetZeroDurationFiles : IRequest<ZeroDurationFilesViewModel>;